✨(backend) create the Referral model
The purpose of building Partaj as an application is to handle
referrals between public agents and offices. We're starting by the
top, creating a somewhat fleshed-out Referral model, which will be
evolving as we figure out our actual needs.

We'll also add dependencies (such as a proper user model) as we go
along.