♻️(backend) add string representation to Referral model
As we were looking at referrals in the admin interface, we noticed
it was not very helpful to have referrals displaying their ID as
an identifier.
We used the model name with a substring of the referral subject
instead as that should be useful wherever we encounter this
string representation.