🐳(terraform) create a Dockerfile for openstack CLI
We'll be deploying to OVH, and need to use object storage there. To
make this easy to handle across environments, we decided to manage
our buckets with terraform.
OVH uses OpenStack and we will therefore need the openstack CLI
tools to interact with it. This dockerfile should make this easy
and reproducible without polluting our host.