🔥(infra) remove Terraform setup for deployment on OVH
Deploying on OVH turned out to be far from painless as it's just
an IaaS and requires us to work through seldom-documented Terraform
and OpenStack features to end up with a non-ideal setup where we'll
have to manually handle our infra.

We're opting for a PaaS offering instead to save time on this. Keep
this history in git to reflect our experiments with deployment
strategies.