🗃(migrations) add a migration to seed the first superuser
We're deploying through a docker container. We want an easy way to
seed the first superuser, without creating duplicates on each deployment,
and without logging in through ssh (as that is not available).

We can use a data migration, which will create the first superuser on
any given deployment. It will only run once as this is a promise of
Django migrations.