🐳(infra) run migrations before starting up gunicorn server
We need to run Django migrations before the gunicorn server is
started on Clever Cloud. This is a basic necessity for Django to
even run.

The most straightforward way to achieve this when we do not have
access to our container as it runs (and we would rather avoid doing
it manually), is to just add it to the CMD for our Dockerfile.