💡(settings) add comment around ALLOWED_HOSTS use
Django needs ALLOWED_HOSTS to be defined with the actual host that
will be part of requests is receives, yet we just set it as None in
the Production settings.

Add a comment to explain that Django will automatically pick up the
related environment variable and use its content to match Hosts
against.