🔧(assets) update settings and urls to serve static files
We want to serve static files with Django in development. Update
settings and urls to enable this for both app statics and global
partaj statics.

This also sets up collectstatic within /data to receive all statics
that need to be available in production. Actually deploying them and
serving them through Clever Cloud will be done later on.