🍱(infra) add WhiteNoise to serve assets in non-dev environments
Django cannot serve the statics itself in non-Development
environments. This is by design as it's mostly not a good idea.

We can use WhiteNoise as a middleware to easily serve our statics
through python (though it delegates to system calls through gunicorn).

This is the only easy way to serve files without much fuss with our
PaaS. If we need to scale our approach up, we will have to add a CDN
later on.