🔧(infra) update settings to add a Staging environment
As we're nearing Partaj's soft launch with beta testers, we want to
add a Staging environment to test our product in a real environment
without breaking the production for no good reason.

All settings are similar to Production as Staging is intended to
run on the same PaaS under the same parameters.