🐳(build) add environment variables to Dockerfile
We want to run some Django-related deployment tasks before launching
the application server. This is not possible for migrate as the
database connection is not available during image build. We can
however move others commands, such as collectstatic, to RUN calls
if we make the relevant environment variables available to the
Python process.