🐳(build) add a compilemessage step to Docker build
In development, we extract translation files and complete them with
localized strings. We can commit those as they are human-readable and
we intend to manually edit them in the repo for now.

While running in production, Django uses binary translation files
that it is not useful to commit to the repository as they are
deterministically generated from our .po translation files. We
can instead just regenerate them as part of our deployment build.