✨(referral) update the referral form view to allow file uploads
Since we linked attached files to referrals through another model
with a OneToMany relationship, we need to handle them manually in
the view rather than just delegate to the ModelForm.
We're adding in a file input, using the browser's default UI for now,
and creating ReferralAttachment instances in the view handler for
each uploaded file.