✨(attachments) update attachment names in storage
We should not upload referral attachment files straight to the
bucket with their own filename, as this can create conflict and
cause older files to be overwritten if something new is updloaded
with the same name.

We're prefixing filenames as uploaded by users with the UUID for the
referral attachment. As those are unique, we should avoid any risk
of conflicts between referral attachment. The only case where a file
would be overwritten is if a user uploads a new file with the same
name on the same referral attachment object. This would then be a
feature, not a bug.