✨(attachments) only serve file attachments to authenticated users
Referral attachments are not meant to be publicly accessible. We can
prevent that by just uploading them with a "private" acl by default,
using the django s3 storage settings.

The only remaining challenge is to serve them only to authorized users.
For now we're focusing on just validating user authentication. We
added a Django view that checks the user's auth and serves the file
linked to the referral attachment.