✨(referral) add information to our Referral model
We'll be launching without an authentication system for our users
during the first part of our beta. We therefore need clearer
identifying information, including phone numbers and email adresses
to be able to communicate with referral requesters.

We also added a "question" field meant to hold the proper subject
of the referral, where "subject" should be a topic that's just helpful
to point the referral to the right expert.