✨(notifications) add a transactional email for referral creation
Users need to be notified by email that their referral has been
created. For now, this email only contains plain text information
but we'll enrich it as needed.

It lives on our email provider's GUI and admin tools, and we stitch
it with our app through the template ID in the settings. This
approach lets us move fast and is easy to change if/when we need to.