✨(referral) update the referral form
As we remove the 200-char limit for the question field on the
Referral model, we had to reflect this change in the form control
used to input the question. We just replaced the text input with
a textarea.

We also noted that there were mistakes in the labelling of some
forms elements and took this opportunity to fix them.