💬(referral) update wordings in the referral form
As we streamlined the referral form, we noticed some model field
help texts were not very helpful and were just redundant with the
field names. We replaced them with something different and more
helpful directly in the form.

We also added a small disclaimer at the end of the form that users
could not, at this time, change their referral after they have
created it.