💥(database) re-do migrations from scratch
Django does not support adding a custom User model after initial
migrations have happened. There is a convoluted process, including
manually running SQL on the database and faking migrations, that
makes is possible to sidestep this issue.

Given the state of our application, it seems simpler, quicker and
safer to allow ourselves to drop the database and reload our data.

Therefore, we re-did the migrations from scratch and will run them
on brand-new databases when we deploy.