🐛(admin) set default referral attachment name in admin
In the django admin site, referral attachments are created directly
without ever going through our referral creation view. They're hence
missing the part that defaults the referral attachment "name" field
to the file name (without extension) if it is not filled by the user.

We can subclass `ModelForm` to simply replicate this behavior in the
admin site.