💥(backend) use incrementing integers instead of uuids for referrals
Django does not support using an AutoField on a model if it is not
a primary key, due to limitations in some database software. We want
unique integers for referrals to give us human-redable ids.

As we already have to break the database to change the user model,
we can just replace our uuids with auto-incrementing integers on
the referral model.