✨(models) add a linked user field on the Referral model
As we recently added login/signup through a CAS/SAML identity
provider, we now have user accounts. This means we can link referrals
to the user that created them in the first place.

This will enable us to more easily manage access rights, referrals and
responses as well as notifications moving forward.