🐛(auth) redirect users to index after log out
When users log out, we need to redirect them to the index, which
is accessible to anonymous users. This is useful as it prevents them
from being re-authenticated automatically through Cerbère, which
would happen if we bounced them back to a view that requires users
to be logged in.

We had to update `django-cas-mb` and cut a new version of it to
get a new setting that enables us to achieve this.