✨(backend) add a logged-out index view
The "new referral" view was our index view. Now that we can log users
in through Cerbère, we decided to put it behind an authentication
barrier.

This means we need some other view to act as a logged out index so
we do not just force login on all users when they first reach Partaj
(due to the interaction between django's auth and django-cas).
It also provides us with a view where we can send back users after
they log out.