🗃(backend) make requester phone & email optional on Referral
Now that we can connect a User instance to Referrals as we create
them, we no longer need to directly collect email and phone number
for the user on the Referral itself.

We're making these fields optional, before removing them entirely
at a later point. This allows us to keep information on existing
Referrals until we can link them back to their user.