✨(form) auto-fill user information on the referral form
As we can log users in with Cerbère and get some information about
them, we don't need to have them manually fill in their name, email
and phone number in the form anymore.

Instead, we can add readonly fields that show what we have already
gotten about them.

We kept the requester name field to allow users to request a referral
to be formally adressed to someone else than the user that is currently
logged in, eg. a préfet, ministre or entire department.