✨(form) update the "referral received" view
Now that referrals are relying on the linked user, we should not
show the Referral "requester_email" field on the confirmation page,
and should instead show the linked user's email address.

Also, as our Referral ids now are integers, meaning they are human
readable, we can use them as case numbers and display them on the
referral received view.