💄(form) lighten the display of user information in the Referral form
We need to display the user information we got from Cerbère on the
Referral form, so the user knows who they're making the referral
as and how they will be contacted by us.

We previously built this with readonly form fields. This took up a
lot of space and could be confusion. We've made this simpler and
lighter.