✨(form) replace freeform subject with predefined topic in referral form
Now that we have topics and units in the database and can link them
with referrals, we can replace our temporary freeform subject field
in the Referral form itself with the more useful predefined topic
field.

Also, we can't make it required in the model due to database constraints
for now, but we can and did make it required in the form itself.