♻️(backend) rework email utils to support multiple emails
Until now, we were only sending one email with Partaj. Everything
was hardcoded there, which made sense.

As we gear up to send more kinds of emails to our users, it is
time we create a utility class to factorize and consolidate all
the relevant data and reuse all the logic we can.