✨(models) add a helper method on Unit to get the organizers
To determine who to send an email to when a unit receives a new
referral, we need to get a list that includes all its admins and
owners (and any other managing roles we might add in the future).

Instead of hardcoding this in the email helper, it is simpler to
add a helper method onto Unit so we can have this list available in
the future should we need it elsewhere in Partaj.