✨(emails) send an email to admins/owners when a unit receives a referral
We just created associations between users, units, topics and referral.
We can start drawing on that to send an email to alert the admins and
owners of a unit when the unit receives a new referral. This is critical
because, as the receivers of a referral, they need a notification to
even know it exists.