✨(unit) create a view where users can view all referrals for a unit
For now, the new unit inbox is a simple table view where members of
a unit can view all referrals for their unit. It will be the starting
point for a whole part of the app that will be dedicated to managing
and responding to the referrals a given unit receives.