✨(nav) add a link to unit inbox on in the app navigation
We just added the first view for the new part of our app that will
enable unit members to manage and respond to referrals. We need
them to be able to access this view easily, therefore we're just
adding a link to the main navigation, only for users who belong
to at least a unit.