🗃(membership) make unit memberships unique for a user & unit
When we created the UnitMembership model, we did not add a constraint
that a user can only have one membership with a given unit. It was
presupposed by our mental model but not reflected in our actual Django
model.

We're adding this constraint as it makes it way easier to reason about
memberships and write code that checks/manages them.