✨(models) create a model to handle referral assignments
We need to support assigning a given member of a unit a referral
linked to that unit to work on.

To have control and be able to keep some basic information on such
assignments, we decided to use a dedicated model to manage them.

We also configured the admin view to manage the model.