♻️(views) create a folder to split views into different files
The `views.py` file was starting to get confusing, with a lot of
different views with different purposes and functions living
together.

As we get a clearer picture of how the app is laid out, we can
split the views into files so related mixins and/or helpers get
an appropriate location.