♻️(base) remove the (current) sr marker in the main app navigation
The '(current)' sr marker was hardcoded on what was then the only
view in Partaj. We're better off removing it as it provides false
information, and adding it back in when we implement the active
functionality in the main navigation both as a visual cue and as
a sr-accessible marker.