✨(models) add a method to get attachment name with file extension
In some contexts we want referral attachments to be listed including
the complete name and the file extension, but not the actual path in
our bucket, which includes an unfriendly uuid.

We can just concatenate the attachment name with a file extension
from the attachment file itself to obtain the desired string.