✨(unit) add a side submenu for the "unit" views
We're starting to flesh out a whole part of Partaj that will deal
with units. We'll need to enable users to access all of these new
views easily.

To that end, we're adding a submenu that now only points to the inbox
but that also hints at the views we'll be adding later on.