🗃(models) add size to referral attachment model
We want to display referral attachment file size when we offer it
to users for download, as a cue for what to expect wrt. download
time and space taken up.

However, accessign this data after the file is uploaded is problematic
as the files are stored in a remote s3-like bucket: it would require
authenticating with the bucket and making a HTTP round-trip.

Instead, we can save the file size on the model and have it handy
whenever we need to display it.