✨(attachments) add attachment file size to the model & unit detail view
We added a size field to our referral attachments to be able to display
attachment file sizes whenever we need to. However, wee first need to
get that file into our model.

To avoid repeating this throughout the code base and in the admin views,
we made an override to the attachment model that adds the file size
before creating the referral attachment object.

We can then easily display it in the unit referral detail view if
available.