♻️(models) add default referral attachment name to save override
We previously handled defaulting referral attachment name to the
linked file name in the view and in the admin form. This duplicated
logic and also left a hole as it did not work when creating an
attachment through other means, eg. an inline in the admin site.

We can instead add it in the referral attachment save override so
we can be sure the work is performed no matter where the attachment
is created. It also avoid duplicating this small bit of logic.