💥(models) remove referral assignment model
We created the referral assignment model as a way to track assignments
and keep information about assignments.

As we reflected on an overall architecture plan for Partaj, we realized
assignments were not particular case: there are lots of events that
we need to keep track of to build a history stream for a Referral, while
also modifying the referral itself to more easily add features there.

We therefore decided to remove our assignment model and instead use an
assignee field on the referral (to be added after this commit) to track
the current status of a referral. The history stream will be managed in
a new event/activity model that is not used to reconstruct the actual
state of a referral at a given time.