🗃(models) remove length limit on referral urgency explanation
Justifying an urgency level is not something that properly fits in
a simple short input text. It can be a long textual explanation requiring
a textarea, and therefore maps more logically to a TextField than a
CharField.

We're simply making this change in the model and reflecting it in
the actual form element we use in the referral creation view.