🗃(models) remove subject field on Referral model
The subject field on Referral was replaced with topic, which is a
ForeignKey on a dedicated model instead of being a free text field.
This is much more useful to sort and manage referrals.

Now that we have this new field, and have ensured data integrity in
all our active databases, we can just remove the subject field
entirely.