🗃(models) add assignments as a ManyToMany on Referral
The main use-case for assignments is to be able to easily get a list
of assignees for a referral and add people to it if the current user
is authorized.

We therefore decided to created assignments as a ManyToMany on
Referral, creating an intermediate model to cleanly add some
relevant information about those assignments.