🙈(frontend) create a folder where Django picks up the frontend build
We need the frontend build to be picked up by Django during
collectstatic so it is served locally and placed in the right location
to be served in live environments.

This directory needs to be gitignored to avoid errors where we commit
frontend build artifacts, but committed to make sure Django's
collectstatic can always find it without fiddling around. Thus the
gitkeep.