🔧(frontend) handle webpack public path for lazy-loaded chunks
We're enabling lazy-loading by default in our app. To do any of
it, we need to tell webpack about the place where the chunks will
be accessible on the live website.

This is where public path comes in, it lets us set an appropriate
default that works out of the box and specify another value server
side that would be picked up by the client.