✨(frontend) create the skeleton of our frontend application
Rather than build an SPA right now, we're planning on incorporating
React-rendered components where user interaction happens and using
server-side hacks makes no sense. We'll progressively extend our
components where and when it makes sense.

To achieve this, we need a skeleton with an index and a root component
that uses React portals to render component to locations on the page
that are tagged with a specially named class.

This skeleton also takes care of wrapping everything in an i18n
provider and loading whatever polyfills we need to run our app.