➕(backend) add django-fsm to our dependencies
Most of a referral's lifecycle can be characterized as a bunch of
state transitions. Instead of weakly managing them through manual
changes on the model, we want to use a dedicated library and model
field to build something more robust.

This should also be helpful to organize code that handles everything
that needs to happen during a transition: sending emails, creating
activity stream objects, modifying the referral itself.