♻️(models) make Referral state field a FSM field
We manage the current state of a referral through a bunch of choices.
This can be done naively with a CharField which is manually changed
whenever we need to. However, if we want to encode our business logic
in a way that is more resilient to errors and exhibits less undefined
behavior, we can instead use a finite state machine.

Here, Django-FSM can help us by providing a field and model method
decorators to codify and handle the transitions for us. The protected
attribute means the field can't just be mutated in-place, preventing
invalid transitions.

We also took this opportunity to use an enum (through the django-friendly
models.TextChoices) to encode the state themselves in a more robust and
easier to use way.