✨(api) open a Referral endpoint to assign the referral to a user
The first UI element we're building on the frontend is a small
widget to enable unit organizers to assign a referral to a member
of their unit.
This should also trigger a state change, depending on the previous
state of the referral.

We're adding an API route and hooking it up to a FSM transition
method on the Referral model.