✨(backend) add a context processor for frontend use
Create a partaj_context context processor and use it to create a
FRONTEND_CONTEXT value. We can then set that value in base.html to
a global variable that will be trivially picked up by our frontend
application.

This is the simplest way to pass global shared content from backend
to frontend.