♻️(unit) rework unit submenu to be horizontal
The side navigation used for the unit views was fine as a general
concept, helping to organize the content for the whole sub-part of
the app. However, it also took up critical horizontal space and
prevented us from using columns in our content.

Switching it to horizontal and on top gives us more room for columns
in the referral inbox, and enables us to create a right-hand side
column in the referral detail view for status information and UI
elements.