♿️(styles) add a utility class for offscreen elements
We want some elements to appear to assistive technology users but
be hidden visually. The main use-case for this is text alternatives
to custom visual elements on the view, such as our new spinner.

For this, we can just use absolute positioning to ensure the text
does not appear on the screen even though it keeps its place in the
DOM.