🔧(frontend) set up the translation workflow for frontend strings
We use react-intl to automate file conversions and generate the
initial POT file for translation work as well as the JSON files
which are loaded by webpack.

We can then translate our strings with translation software such
as POEDIT which enables us to generate PO files from a POT template
and update them with new strings as they come in.