🐳(frontend) update Dockerfile to build and distribute JS build
Now that we actually have frontend components to distribute with
Partaj, we also need to run the frontend build and make sure the
output bundle is distributed along with the app.

To this end, we're adding a node step to our Dockerfile to run the
frontend build, and copying these files into the image we're actually
running at the end.
We also added a package.json script to run webpack in production
mode.