♻️(backend) use settings to get the app location
When we send emails from Partaj, we sometimes need to include a link
back to the application. For that, we need to know where the app is
running (eg. scheme and domain).

We were being "clever" by using the request to give us that (as surely
it contains a way to reach the app). However this forces us to pass
the request through unrelated elements such as models and mailers.

Django has a Site model that could take care of this for us, however
it does not provide a scheme, which we need to build our links. We
therefore decided to add a setting where everything we need can be
specified.