✨(referral) send an email to the assignee for a referral
Assignees need to be notified when a new assignment is created for
them on a referral. We can use the same mechanism as for other
emails, in our Mailer module.

Since we have a transition for assignments on referrals, we can use
it to handle this call to the mailer. It gives us a logical place
for this business logic.