💄(unit) simplify case numbers in unit referral inbox
The unit inbox has a lot of data to display. Case numbers are taking
up a lot of space, even though they are very compact, due to the
length of the column header.

We can replace it with "#", and add the same in front of the case
numbers so users can easily make sense of the data, and we can get
back some valueable screen space. We kept the textual header as an
a11y header, while the new one is hidden, because it makes more
sense there as the visual context is missing to interpret the "#".