🗃(models) add a ReferralAnswer model
Create a model to keep the referral answer. We're using this approach
instead of adding a field to Referral so we can keep information such
as when the referral was created and by whom.

It will also make it easier if/when we need to add more information to
the answer, such as a title, some separate UI elements or attached
files.