✨(admin) show referral answers in the admin site
Create an inline admin class for referral answers so we can show
them in the referral admin view.
For now there's no reason to create a separate admin view and list
view for referral answers as they don't make a lot of sense outside
of their related referrals, and we don't have too much admin work
to do on them either.