✨(frontend) add a referral answer form/display on unit referral detail
After assigning a referral to a member, the next thing a unit needs
to do is to actually provide an answer.

To that end, we're adding a component at the bottom of the page that
either shows the answer if it already exists, or offers a form where
a unit member can provide the referral answer, using our newly opened
API endpoint.