👷(frontend) create a github actions workflow to run frontend checks
We started building Partaj without any continuous integration
pipeline, simply pushing to staging & production environments after
running checks locally. This is fine when starting a greenfield
project, but as we are growing and users are counting on us in
production, we want to run a tighter ship.

Create a workflow to run our frontend jobs when pushing to master
or creating a pull request against master. Start with the one job
we already have, the Typescript build.