✨(models) add a method to unit to get related memberships
We need a simple way to get all members of a unit along with their
membership (to this unit, as opposed to a membership set of all
their memberships).

This is not easy to achieve starting from members, as we would like
to annotate a new field that points to a nested object, which as far
as we know is not supported with Django's ORM expressions. Instead,
we can go the other way around: if we just get the memberships and
select their user, we get everything we need without too much ad
hoc code.