♻️referral) clean up referral state css classes
We started out with an if block in our unit referral inbox to
determine which color class to apply to a referral depending on its
state.

We want something more generic instead, using a method of the class
allows us to write cleaner code and reuse it when we need it.