🔧(backend) set up the project to run backend tests with pytest
We want our tests to run in the same kind of docker-compose
environment we're using for developmentn so we don't have any
version issues to manage and can ensure our environements are as
close as possible.

We're using a bin folder to automate things and simplify what we
have to type when we want to run the tests.