💄(content) respect white space in user-generated content
By default HTML does not respect white-space at all in content,
disregarding linebreaks and spaces/tabs.
This is not what we want for user content: the bare-minimum we're
going for is letting users format their messages with linebreaks.

Using `white-space: pre-wrap` for paragraphs by default should help
us get this behavior everywhere we need it.