♻️(templates) move referral detail content to a partial
We're about to build out a few simple views for requesters where
they can find their referrals, track their state and read them and
their answers.

The layout will be similar to the view for the unit but it will have
significant differences in authorizations and features. To reuse as
much as we can, we'll take advantage of frontend components, and
partials for the backend part.