💄(styles) change white-space class for user content
We previously added a blanket rule for `white-space: pre-wrap` to
all paragraphs in Partaj. This was a little bit ambitious as we had
missed some paragraphs which are not user generated content and do
not benefit from taking white space into account.

We should use a class instead to target this only to user content.