✨(requester) add a view where users can see all referrals they've created
Now that we have gotten started on a more substantial unit part of
Partaj, we need to build some basic views for requesters: they need
to be able to find their referrals in Partaj and check their status.

The first step is to add a list view where they can see the referrals
they've created until now.