✅(tests) add a Deferred util to more easily use promises in tests
We want to control how and when promises are resolved in our unit
tests. This can be useful to eg. manage a response to a fetch with
proper act() wrapping and test intermediate state(s) between promise
resolution(s).

This is clunky to do with just the Promise constructor as it does
not have a way to resolve a promise from outside of its own scope.
Deferred gives us a clean and easy way to get this.