♻️(frontend) add a hook to use the current user
We often need to know who the current user is so we can manage
authorization and see what UI elements they can see and what
actions they can take.

To avoid repeating this, we're creating a custom hook that makes
it trivial for any component to get the current user.