✨(frontend) add a utility method for unit member permissions
In addition to the existing method that checks if a user is an
organizer for a given unit, we need one that checks if they're
a member of any kind of that unit.

We took this opportunity to add tests for the existing unit
utility methods, which revealed a small issue with the handling
of null.