✨(frontend) do not show the referral answer form to non-members of unit
Only members of the related unit can answer a referral. This is already
enforced through permissions in the API endpoint. We should therefore
not even show the form to other users as it can be confusing for them,
and can only lead to errors if they try to use it.