✨(referral) generate activity in the referral transition methods
As we built referrals around the transition methods and use them
to both create related objects and send emails, they are the
natural place to also create the activity objects.

We took this opportunity to create an initial transition, called
send. For now, it does not affect the state, but it will make it
very easy for us to have draft referrals later on. It's also a
better place for mailer calls as well as activity creation.