✨(api) add referral activities on referral objects in the API
We want referral activities to be available on the referral as
they are the way we primarily structure referral detail pages.
We need to build them in the frontend as opposed to Django
templates as we want to show them instantly when unit members
create new events, and avoid relying on javascript page refereshing.