🏷️(frontend) add typings for referral activity
We recently added a new ReferralActivity model to represent events
that happen around a referral. These are available on referrals in
the API as the frontend will need to display them in the referral
detail view.

We can add them through keyed types using the verb to discriminate
between various activity kinds with various objects on the
item_content_object field.