♻️(referral) name the reverse relation for attachments on referral
When we implemented referral attachments, we did not add a name
for the reverse relation. As we tried to add attachments them to
our referral serializer, this required us to write some imperative
glue code to get the attachments.

We should instead just give a name to the reverse relation. This
will also make it easier to get attachments linked to a referral from
anywhere.