♻️(frontend) split <ReferralActivityIndicator /> looks from content
We want to reuse the <ReferralActivityIndicator /> look for ongoing
activity (for example, to top the referral answer form). To reuse the
structure and styles, we're splitting the content (essentially two
separate JSX elements building text) from the structure of the
indicator itself.

We can then reuse it by passing it what should be shown in the top
and bottom line.