✨(frontend) integrate referral answer form with the new referral detail
As we implement the referral detail view redesign, we want the referral
answer form to integrate seamlessly into the new design.

We needed to manage the answer form display state from higher in the
component tree as it coordinates a button from <ReferralDetailContent />
with a component to display in <ReferralDetail />.

We also changed the look of the form for the redesign and took this
opportunity to replace bootstrap components with tailwind classes.