💄(frontend) fix broken bootstrap styles
We want to move from bootstrap to tailwind to power our UI styles.
There is some conflict there as tailwind has a CSS reset that
erases bootstrap styles and bootstrap has !important flags on
some utility classes that have the same name as tailwind styles.

We'll remove bootstrap entirely very soon. In the meantime, we're
fixing whatever is broken throughout the app with the addition of
tailwind.