🔥(templates) delete the referral detail partial
Now that all the code for the frontend of the referral details view
has been moved to our actual React frontend, there is no point in keeping
a partial that does not reuse any code.

We can remove the partial and paste the call to the frontend in both
versions of the referral details view.