🐛(assignment) do not show assignment button on invalid states
Users can only be assigned on referrals that are in some valid states,
including RECEIVED and ASSIGNED, which are enforced by the FSM in the
backend and cause API requests to fail if the transition is attempted
from an invalid state.

Instead of showing the button and failing silently for users, we need
to hide the button when the assignment is not possible anymore.