💄(frontend) add a componentized btn class using tailwind
The tailwind modus operandi is to mostly avoid creating classes
for components, instead reusing the actual HTML generated and
utility classes.

This is however not really appropriate for buttons which are peppered
all over the app, in both backend and frontend. This is one of the
cases where we can make "component classes" of our own.