➕(backend) add and setup django-impersonate
We are encountering hard to explain bugs that only some of our
users are seeing. We have tracked them down to authentication
issues.

To help track them down (and do that more easily in the future,
should the need arise), we're adding django-impersonate, which
enables us to see and use the app as if we were someone else.