✨(frontend) add a hook to detect clicks outside a given element
We're adding React powered dropdown menu to <ReferralDetailAssignment />.
To make it function as users expect it, it should close itself when
they click anywhere else on the page.

We can write a simple hook using refs and a click handler on document
(+ cleanup) to detect outside clicks and call a handler at this point.