🍎(styles) fix referral list views on safari
The way links are stretched breaks Safari as the table-row which
we want to be completely covered by the link cannot be a containing
object in Safari.

We dont want to change elements or layout either as it could
compromise accessibility. For now, we're just disabling the link
stretching in Safari until we can re-implement these tables as
role="grid" and get rid of the problem altogether.