🗃(models) adjust models to support the unassign transition
ReferralActivity needs a new possible Verb to support the
"unassign" transition.

We also needed to remove "protected=True" from our Django-FSM
state field on Referral. This is not because we want to assign to
it directly, but because its current implementation is broken: it
breaks `refresh_from_db()` which is a commonly used method on django
model instances and is needed to update the referral after deleting
an assignment.