✨(api) add an action route on referral to unassign someone
This new route is the reverse of assign, that removes an assignment
from a member who was assigned on a referral. It also uses a transition
action on the Referral model.

It does not however remove the referral activity pertaining to the
initial assignment. Instead, it adds a new activity marking the removal
of the assignment, keeping a complete history of the referral.