✨(frontend) improve our <ReferralDetailAssignment /> component
We wanted our assignment component to take up less space, and still
offer a dropdown menu that allows to assign any member of the related
unit. We also took this opportunity to add unassignment functionality
to the same dropdown menu.

We used <fieldsets> for the buttons to make them accessible and added
test cases to try and go through the possible assignment and
unassignment scenarios.

We had to split <ReferralMemberAssignmentButton /> to a separate component
to more easily represent its states through an XState machine.